put logger and enviroments in dif files. start mess with the layout
